feat: moves auth operations to a nested Auth struct
