+++ src/passage_flex/test.rs
@@ -1,3 +1,5 @@
+use crate::Error;
+
 use super::*;
 use mockito::{Matcher, Mock, Server, ServerGuard};
 
@@ -222,6 +224,7 @@ async fn test_create_register_transaction() {
         .await;
 
     let transaction_id = passage_flex
+        .auth
         .create_register_transaction("test".to_string(), "test".to_string())
         .await
         .unwrap();
@@ -244,6 +247,7 @@ async fn test_create_authenticate_transaction() {
         .await;
 
     let transaction_id = passage_flex
+        .auth
         .create_authenticate_transaction("test".to_string())
         .await
         .unwrap();
@@ -260,7 +264,7 @@ async fn test_verify_nonce() {
     .with_body(r#"{"error": "Could not verify nonce: nonce is invalid, expired, or cannot be found","code": "invalid_nonce"}"#)
     .create_async().await;
 
-    let invalid_result = passage_flex.verify_nonce("invalid".to_string()).await;
+    let invalid_result = passage_flex.auth.verify_nonce("invalid".to_string()).await;
     assert!(invalid_result.is_err());
     assert!(matches!(invalid_result, Err(Error::InvalidNonce)));
     m1.assert_async().await;
@@ -277,6 +281,7 @@ async fn test_verify_nonce() {
         .await;
 
     let external_id = passage_flex
+        .auth
         .verify_nonce("valid".to_string())
         .await
         .unwrap();
@@ -348,101 +353,4 @@ async fn test_revoke_device() {
     m2.assert_async().await;
     m3.assert_async().await;
     assert!(result.is_ok());
-}
-
-#[tokio::test]
-async fn test_get_app() {
-    let mut server = Server::new_async().await;
-    let api_key = "test_api_key";
-
-    let app_id = "invalid";
-    let mut invalid_passage = PassageFlex::new(app_id.to_string(), api_key.to_string());
-    invalid_passage.set_server_url(server.url());
-    let m1 = server
-        .mock("GET", "/v1/apps/invalid/")
-        .with_status(401)
-        .with_body(r#"{"error": "Invalid access token","code": "invalid_access_token"}"#)
-        .create_async()
-        .await;
-    let invalid_result = invalid_passage.get_app().await;
-    m1.assert_async().await;
-    assert!(invalid_result.is_err());
-    assert!(matches!(invalid_result, Err(Error::InvalidAccessToken)));
-
-    let app_id = "test_app_id";
-    let mut passage = PassageFlex::new(app_id.to_string(), api_key.to_string());
-    passage.set_server_url(server.url());
-    let m2 = server
-        .mock("GET", "/v1/apps/test_app_id/")
-        .with_status(200)
-        .with_body(
-            r#"{
-        "app": {
-            "additional_auth_origins": [],
-            "allowed_callback_urls": [],
-            "allowed_identifier": "external",
-            "allowed_logout_urls": [],
-            "application_login_uri": "",
-            "auth_fallback_method": "none",
-            "auth_fallback_method_ttl": 300,
-            "auth_methods": {
-                "magic_link": {
-                    "enabled": false,
-                    "ttl": 300,
-                    "ttl_display_unit": "s"
-                },
-                "otp": {
-                    "enabled": false,
-                    "ttl": 300,
-                    "ttl_display_unit": "s"
-                },
-                "passkeys": {
-                    "enabled": true
-                }
-            },
-            "auth_origin": "https://auth.test.com",
-            "auto_theme_enabled": true,
-            "created_at": "2021-01-01T00:00:00Z",
-            "default_language": "en",
-            "element_customization": {},
-            "element_customization_dark": {},
-            "hosted": false,
-            "hosted_subdomain": "test",
-            "hosted_theme": "auto",
-            "id": "test_app_id",
-            "id_token_lifetime": 600,
-            "layouts": {
-                "profile": [],
-                "registration": []
-            },
-            "login_url": "/",
-            "name": "Test App",
-            "passage_branding": true,
-            "profile_management": false,
-            "public_signup": false,
-            "redirect_url": "/",
-            "refresh_absolute_lifetime": 2592000,
-            "refresh_enabled": false,
-            "refresh_inactivity_lifetime": 432000,
-            "require_email_verification": false,
-            "require_identifier_verification": false,
-            "required_identifier": "external",
-            "role": "",
-            "rsa_public_key": "",
-            "secret": null,
-            "session_timeout_length": 5,
-            "technologies": [],
-            "type": "flex",
-            "user_metadata_schema": []
-        }
-    }"#,
-        )
-        .create_async()
-        .await;
-
-    let app_info = passage.get_app().await.unwrap();
-    m2.assert_async().await;
-    assert_eq!(app_info.id, "test_app_id");
-    assert_eq!(app_info.name, "Test App");
-    assert_eq!(app_info.auth_origin, "https://auth.test.com");
 }
